+++ nagra/__init__.py
@@ -3,4 +3,4 @@ from .table import Table
 from .schema import Schema
 from .transaction import Transaction
 
-__version__ = "0.1.2"
+__version__ = "0.2.0"