+++ nagra/table.py
@@ -50,27 +50,33 @@ from nagra.transaction import Transaction
 from nagra.exceptions import IncorrectTable
 
 
-_TYPE_ALIAS = {
-    "str": "str",
-    "varchar": "str",
-    "character varying": "str",
-    "text": "str",
-    "int": "int",
-    "integer": "int",
-    "bigint": "bigint",
-    "float": "float",
-    "double precision": "float",
-    "timestamp": "timestamp",
-    "timestamp without time zone": "timestamp",
-    "timestamptz": "timestamptz",
-    "timestamp with time zone": "timestamptz",
-    "date": "date",
-    "bool": "bool",
-    "boolean": "bool",
-    "uuid": "uuid",
-    "json": "json",
-    "blob": "blob",  # TODO ADD TEST
-    "bytea": "blob",
+UNSET = object()
+
+
+# Intentionally sorted by reverse lenght to help type hint detection, see Schema._db_columns
+_TYPE_ALIAS ={
+    'timestamp without time zone': 'timestamp',
+    'timestamp with time zone': 'timestamptz',
+    'character varying': 'str',
+    'double precision': 'float',
+    'timestamptz': 'timestamptz',
+    'timestamp': 'timestamp',
+    'datetime': 'timestamp',
+    'boolean': 'bool',
+    'integer': 'int',
+    'numeric': 'float',
+    'varchar': 'str',
+    'bigint': 'bigint',
+    'bytea': 'blob',
+    'float': 'float',
+    'blob': 'blob',
+    'bool': 'bool',
+    'date': 'date',
+    'json': 'json',
+    'text': 'str',
+    'uuid': 'uuid',
+    'int': 'int',
+    'str': 'str',
 }
 
 _DB_TYPE = {
@@ -116,7 +122,7 @@ class Column:
             self.dtype = dtype.strip()
             self.dims = ""
         try:
-            self.dtype = _TYPE_ALIAS[dtype.strip()]
+            self.dtype = _TYPE_ALIAS[dtype.strip().lower()]
         except KeyError:
             raise ValueError(f"Type '{dtype}' not supported (for column '{name}')")
 
@@ -169,7 +175,7 @@ class Table:
         self.not_null = set(self.natural_key) | set(not_null or [])
         self.one2many = one2many or {}
         self.default = default or {}
-        self.primary_key = primary_key or "id"
+        self.primary_key = "id" if primary_key is None else primary_key
         self.schema = schema
 
         # Detect malformed fk definitions